Preparation for multi file upload (part 2)
Every time a button is clicked, the corresponding message element is generated through jquery. Later when multiple files are uploaded, the system would be able to show the message for each file. Currently, uploading is expected to be done once. Meaning user would not be able to first select x number of files, then select 1 more.