Input now only accepts pdf + fixed minor bug
Input type file now has an accept parameter, which tells the user's OS to only allow certain filetypes. This is not a sure-fire way to limit the filetype being uploaded, hence the older checkings are still in place.

A bug was found and fixed where if the blockchain network is not connected, an error message did not pop up.